UHF-11186: Moved the CKEditor initialize method to test base.
